+++ packages/sdk/src/cli/apply/services/resolver.ts
@@ -409,7 +409,6 @@ function processResolver(
     operationType: resolver.operation,
     response,
     pipelines,
-    postHook: { expr: "({ ...context.pipeline.body });" },
     publishExecutionEvents: executorUsedResolvers.has(resolver.name),
   };
 }